+++ core-java/src/main/java/com/baeldung/enums/Pizza.java
