+++ core-java/src/main/java/com/baeldung/enums/PizzaDeliveryStrategy.java
