+++ core-java/src/main/java/com/baeldung/enums/README.md
